+++ src/components/DarkThemeToggle/DarkThemeToggle.tsx
@@ -1,10 +1,8 @@
 import type { ComponentProps, FC } from 'react';
-import { useContext } from 'react';
 import { HiMoon, HiSun } from 'react-icons/hi';
 import { twMerge } from 'tailwind-merge';
 import type { DeepPartial } from '../../';
 import { useTheme } from '../../';
-import { ThemeContext } from '../../components/Flowbite/ThemeContext';
 import { mergeDeep } from '../../helpers/merge-deep';
 
 export interface FlowbiteDarkThemeToggleTheme {
@@ -31,13 +29,13 @@ export const DarkThemeToggle: FC<DarkThemeToggleProps> = ({
 }) => {
   const theme = mergeDeep(useTheme().theme.darkThemeToggle, customTheme);
 
-  const { mode, toggleMode } = useContext(ThemeContext);
+  const { mode, toggleMode } = useTheme();
 
   return (
     <button
       aria-label="Toggle dark mode"
       data-testid="dark-theme-toggle"
-      onClick={toggleMode}
+      onClick={() => toggleMode?.()}
       type="button"
       className={twMerge(theme.root.base, className)}
       {...props}