+++ src/components/Select/Select.stories.tsx
@@ -1,4 +1,5 @@
-import type { Meta, Story } from '@storybook/react';
+import type { Meta, StoryFn } from '@storybook/react';
+import { BsFlagFill } from 'react-icons/bs';
 import type { SelectProps } from './Select';
 import { Select } from './Select';
 
@@ -7,12 +8,26 @@ export default {
   component: Select,
 } as Meta;
 
-const Template: Story<SelectProps> = (args) => <Select {...args} />;
+const Template: StoryFn<SelectProps> = (args) => <Select {...args} />;
 
 export const DefaultSelect = Template.bind({});
 DefaultSelect.storyName = 'Select';
 DefaultSelect.args = {
   id: 'countries',
+  children: (
+    <>
+      <option>United States</option>
+      <option>Canada</option>
+      <option>France</option>
+      <option>Germany</option>
+    </>
+  ),
+};
+
+export const WithIcon = Template.bind({});
+WithIcon.args = {
+  id: 'countries',
+  icon: BsFlagFill,
   children: (
     <>
       <option>United States</option>