+++ src/components/ToggleSwitch/ToggleSwitch.stories.tsx
@@ -1,4 +1,4 @@
-import type { Meta, Story } from '@storybook/react';
+import type { Meta, StoryFn } from '@storybook/react';
 import { useState } from 'react';
 import { theme } from '../../';
 import type { ToggleSwitchProps } from './ToggleSwitch';
@@ -11,7 +11,7 @@ export default {
   component: ToggleSwitch,
 } as Meta;
 
-const Template: Story<ToggleSwitchProps> = ({ checked, ...args }) => {
+const Template: StoryFn<ToggleSwitchProps> = ({ checked, ...args }) => {
   const [switchChecked, setSwitchChecked] = useState(checked);
 
   const handleChange = () => {
@@ -32,4 +32,25 @@ DefaultToggleSwitch.argTypes = {
       options: [...colors],
     },
   },
+};
+
+export const SmallToggleSwitch = Template.bind({});
+SmallToggleSwitch.storyName = 'Small Toggle switch';
+SmallToggleSwitch.args = {
+  sizing: 'sm',
+  label: 'small toggle switch',
+};
+
+export const MediumToggleSwitch = Template.bind({});
+MediumToggleSwitch.storyName = 'Medium Toggle switch';
+MediumToggleSwitch.args = {
+  sizing: 'md',
+  label: 'default toggle switch',
+};
+
+export const LargeToggleSwitch = Template.bind({});
+LargeToggleSwitch.storyName = 'Large Toggle switch';
+LargeToggleSwitch.args = {
+  sizing: 'lg',
+  label: 'large toggle switch',
 };