+++ src/components/ToggleSwitch/ToggleSwitch.tsx
@@ -1,7 +1,7 @@
-import type { ComponentProps, FC, KeyboardEvent, MouseEvent } from 'react';
+import type { ComponentProps, FC, KeyboardEvent } from 'react';
 import { useId } from 'react';
 import { twMerge } from 'tailwind-merge';
-import type { DeepPartial, FlowbiteBoolean, FlowbiteColors } from '../../';
+import type { DeepPartial, FlowbiteBoolean, FlowbiteColors, FlowbiteTextInputSizes } from '../../';
 import { useTheme } from '../../';
 import { mergeDeep } from '../../helpers/merge-deep';
 
@@ -18,6 +18,7 @@ export interface FlowbiteToggleSwitchRootTheme {
 
 export interface FlowbiteToggleSwitchToggleTheme {
   base: string;
+  sizes: FlowbiteTextInputSizes;
   checked: FlowbiteBoolean & {
     color: FlowbiteColors;
   };
@@ -26,7 +27,8 @@ export interface FlowbiteToggleSwitchToggleTheme {
 export type ToggleSwitchProps = Omit<ComponentProps<'button'>, 'onChange'> & {
   checked: boolean;
   color?: keyof FlowbiteColors;
-  label: string;
+  sizing?: keyof FlowbiteTextInputSizes;
+  label?: string;
   onChange: (checked: boolean) => void;
   theme?: DeepPartial<FlowbiteToggleSwitchTheme>;
 };
@@ -35,6 +37,7 @@ export const ToggleSwitch: FC<ToggleSwitchProps> = ({
   checked,
   className,
   color = 'blue',
+  sizing = 'md',
   disabled,
   label,
   name,
@@ -47,25 +50,28 @@ export const ToggleSwitch: FC<ToggleSwitchProps> = ({
 
   const toggle = (): void => onChange(!checked);
 
-  const handleClick = (event: MouseEvent<HTMLButtonElement>): void => {
-    event.preventDefault();
+  const handleClick = (): void => {
     toggle();
   };
 
-  const handleKeyPress = (event: KeyboardEvent<HTMLButtonElement>): void => {
-    event.preventDefault();
+  const handleOnKeyDown = (event: KeyboardEvent<HTMLButtonElement>): void => {
+    if (event.code == 'Enter') {
+      event.preventDefault();
+    }
   };
 
   return (
     <>
-      {name && checked && <input checked={checked} hidden name={name} readOnly type="checkbox" className="sr-only" />}
+      {name && checked ? (
+        <input checked={checked} hidden name={name} readOnly type="checkbox" className="sr-only" />
+      ) : null}
       <button
         aria-checked={checked}
         aria-labelledby={`${id}-flowbite-toggleswitch-label`}
         disabled={disabled}
         id={`${id}-flowbite-toggleswitch`}
         onClick={handleClick}
-        onKeyPress={handleKeyPress}
+        onKeyDown={handleOnKeyDown}
         role="switch"
         tabIndex={0}
         type="button"
@@ -77,16 +83,19 @@ export const ToggleSwitch: FC<ToggleSwitchProps> = ({
           className={twMerge(
             theme.toggle.base,
             theme.toggle.checked[checked ? 'on' : 'off'],
-            !disabled && checked && theme.toggle.checked.color[color],
+            checked && theme.toggle.checked.color[color],
+            theme.toggle.sizes[sizing],
           )}
         />
-        <span
-          data-testid="flowbite-toggleswitch-label"
-          id={`${id}-flowbite-toggleswitch-label`}
-          className={theme.root.label}
-        >
-          {label}
-        </span>
+        {label?.length ? (
+          <span
+            data-testid="flowbite-toggleswitch-label"
+            id={`${id}-flowbite-toggleswitch-label`}
+            className={theme.root.label}
+          >
+            {label}
+          </span>
+        ) : null}
       </button>
     </>
   );