+++ src/theme.ts
@@ -3,16 +3,20 @@ import { accordionTheme } from './components/Accordion/theme';
 import { alertTheme } from './components/Alert/theme';
 import { avatarTheme } from './components/Avatar/theme';
 import { badgeTheme } from './components/Badge/theme';
+import { blockquoteTheme } from './components/Blockquote/theme';
 import { breadcrumbTheme } from './components/Breadcrumb/theme';
 import { buttonGroupTheme, buttonTheme } from './components/Button/theme';
 import { cardTheme } from './components/Card/theme';
 import { carouselTheme } from './components/Carousel/theme';
 import { checkboxTheme } from './components/Checkbox/theme';
 import { darkThemeToggleTheme } from './components/DarkThemeToggle/theme';
+import { datePickerTheme } from './components/Datepicker/theme';
 import { dropdownTheme } from './components/Dropdown/theme';
 import { fileInputTheme } from './components/FileInput/theme';
+import { floatingLabelTheme } from './components/FloatingLabel/theme';
 import { footerTheme } from './components/Footer/theme';
 import { helperTextTheme } from './components/HelperText/theme';
+import { kbdTheme } from './components/Kbd/theme';
 import { labelTheme } from './components/Label/theme';
 import { listGroupTheme } from './components/ListGroup/theme';
 import { modalTheme } from './components/Modal/theme';
@@ -39,17 +43,21 @@ export const theme: FlowbiteTheme = {
   alert: alertTheme,
   avatar: avatarTheme,
   badge: badgeTheme,
+  blockquote: blockquoteTheme,
   breadcrumb: breadcrumbTheme,
   button: buttonTheme,
   buttonGroup: buttonGroupTheme,
   card: cardTheme,
   carousel: carouselTheme,
   checkbox: checkboxTheme,
+  datepicker: datePickerTheme,
   darkThemeToggle: darkThemeToggleTheme,
   dropdown: dropdownTheme,
   fileInput: fileInputTheme,
+  floatingLabel: floatingLabelTheme,
   footer: footerTheme,
   helperText: helperTextTheme,
+  kbd: kbdTheme,
   label: labelTheme,
   listGroup: listGroupTheme,
   modal: modalTheme,